Refactor submission creation in submit.go and test.go
This commit refactors the submission creation in submit.go and test.go files. It adds a new parameter command to the CreateSubmission function in
codecrafters_client.go file. The command parameter is used to specify the type of submission, either "submit" or "test". This change allows for better code
organization and separation of concerns.

Commit message: "Refactor submission creation in submit.go and test.go"